Upload and downloading photo from firebase working... changes to ViewQRCode QRCodeEditor and deleted TakePhoto act
