+++ QRGameTeam15/app/src/main/java/com/example/qrgameteam15/QRCodeEditor.java
@@ -1,8 +1,13 @@
 package com.example.qrgameteam15;
 
+import androidx.activity.result.ActivityResult;
+import androidx.activity.result.ActivityResultCallback;
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
+import androidx.core.content.FileProvider;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
@@ -11,13 +16,17 @@ import android.content.pm.PackageManager;
 import android.location.Address;
 import android.location.Geocoder;
 import android.location.Location;
+import android.net.Uri;
 import android.os.Bundle;
+import android.os.Environment;
 import android.os.Parcelable;
+import android.provider.MediaStore;
 import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -31,12 +40,18 @@ import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.firestore.CollectionReference;
 import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+import com.google.firebase.storage.UploadTask;
 
+import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 
@@ -63,6 +78,12 @@ public class QRCodeEditor extends AppCompatActivity {
     FirebaseFirestore db;
     CollectionReference collectionReference;
 
+    private ImageView imageView;
+    private String currentPhotoPath;
+    private String filename;
+    Uri imageUri = null;
+    private FirebaseStorage storage;
+    private StorageReference storageReference;
     private QRCode QR;
 
     /**
@@ -221,9 +242,135 @@ public class QRCodeEditor extends AppCompatActivity {
 
     }
     public void addPhotos(View view) {
-        Intent takePhotoIntent = new Intent(getApplicationContext(), TakePhoto.class );
-        takePhotoIntent.putExtra("QRCodeFromEditor", (Parcelable) QR);
-        startActivity(takePhotoIntent);
+        //Intent takePhotoIntent = new Intent(getApplicationContext(), TakePhoto.class );
+        //takePhotoIntent.putExtra("QRCodeFromEditor", (Parcelable) QR);
+        //startActivity(takePhotoIntent);
+        storage = FirebaseStorage.getInstance();
+        storageReference = storage.getReference();
+
+        //EL Start - added 20220321, need testing
+        db = FirebaseFirestore.getInstance();
+        collectionReference = db.collection("Players");
+        // Not using right now because we get the last QR code in firebase - el & manny
+        //QR = (QRCode) getIntent().getParcelableExtra("QRCodeFromEditor");
+        //EL End - added 20220321, need testing
+        Intent takePictureIntent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
+
+        // Create the File where the photo should go
+        File photoFile = null;
+        try {
+            photoFile = createImageFile();
+        } catch (IOException ex) {
+            // Error occurred while creating the File
+            Toast.makeText(QRCodeEditor.this, ex.getMessage(), Toast.LENGTH_SHORT).show();
+        }
+        // Continue only if the File was successfully created
+        if (photoFile != null) {
+            Uri photoURI = FileProvider.getUriForFile(QRCodeEditor.this,
+                    "com.example.android.fileprovider",
+                    photoFile);
+            takePictureIntent.putExtra(MediaStore.EXTRA_OUTPUT, photoURI);
+
+            //startActivityForResult deprecated, use activityResult Launcher instead
+//            startActivityForResult(takePictureIntent, REQUEST_IMAGE_CAPTURE);
+            activityResultLauncher.launch(takePictureIntent);
+
+        }
     }
+
+    ActivityResultLauncher<Intent> activityResultLauncher = registerForActivityResult(
+            new ActivityResultContracts.StartActivityForResult(),
+            new ActivityResultCallback<ActivityResult>() {
+
+                @Override
+                public void onActivityResult(ActivityResult result) {
+                    if(result.getResultCode() == RESULT_OK){
+
+                        File file = new File(currentPhotoPath);
+                        imageUri = Uri.fromFile(file);
+                        filename = file.getName();
+                        uploadImage();
+                        //EL Start - added 20220321, need testing
+                        addPhotoToQRCode();
+                        //EL End - added 20220321, need testing
+                        //Intent intent = new Intent(getApplicationContext(), UserMenu.class);
+                        //intent.putExtra("userMenu_session22", (String) null);
+                        //startActivity(intent);
+                    }
+                }
+            }
+    );
+        /**
+         * Funcation creates a file to store the photo in full-size
+         * @return
+         * @throws IOException
+         */
+        private File createImageFile() throws IOException {
+            // Create an image file name
+            String timeStamp = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
+            String imageFileName = timeStamp + "_";
+            File storageDir = getExternalFilesDir(Environment.DIRECTORY_PICTURES);
+            File image = File.createTempFile(
+                    imageFileName,  /* prefix */
+                    ".jpg",         /* suffix */
+                    storageDir      /* directory */
+            );
+
+            // Save a file: path for use with ACTION_VIEW intents
+            currentPhotoPath = image.getAbsolutePath();
+            return image;
+        }
+
+
+        private void uploadImage(){
+            String username = SingletonPlayer.player.getUsername();
+//        StorageReference imageRef = storageReference.child("images_em/" + fileName);
+            StorageReference imageRef = storageReference.child(username + "/" + filename);
+
+
+            imageRef.putFile(imageUri)
+                    .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
+                        @Override
+                        public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
+                            Toast.makeText(QRCodeEditor.this, "Upload successful.", Toast.LENGTH_SHORT).show();
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                            // Handle unsuccessful uploads
+                            Toast.makeText(QRCodeEditor.this, "Failed to upload" + e.getMessage(), Toast.LENGTH_SHORT).show();
+                        }
+                    });
+        }
+
+        //EL Start - added 20220321, need testing
+        public void addPhotoToQRCode() {
+            int lengthQRCode = SingletonPlayer.player.qrCodes.size();
+            QRCode qrCode = SingletonPlayer.player.qrCodes.get(lengthQRCode-1);
+            qrCode.setImageIDString(filename);
+            qrCode.setHasPhoto(true);
+            SingletonPlayer.player.qrCodes.set(lengthQRCode-1, qrCode);
+            String TAG = "add photo QR working";
+            collectionReference
+                    .document(SingletonPlayer.player.getUsername())
+                    .set(SingletonPlayer.player)
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void unused) {
+                            Log.d(TAG,"message");
+                            Toast.makeText(QRCodeEditor.this, "Add to QRCode successful.", Toast.LENGTH_SHORT).show();
+                        }
+                    })
+                    .addOnFailureListener(new OnFailureListener() {
+                        @Override
+                        public void onFailure(@NonNull Exception e) {
+                            Log.e("MYAPP", "exception: " + e.getMessage());
+                            Log.e("MYAPP", "exception: " + e.toString());
+                            Toast.makeText(QRCodeEditor.this, "Add to QRCode FAILED.", Toast.LENGTH_SHORT).show();
+                        }
+                    });
+
+        }
 
 }
