+++ QRGameTeam15/app/src/main/java/com/example/qrgameteam15/ViewQRCode.java
@@ -1,17 +1,28 @@
 package com.example.qrgameteam15;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.ImageView;
 import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.storage.FileDownloadTask;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 
 public class ViewQRCode extends AppCompatActivity {
@@ -20,6 +31,7 @@ public class ViewQRCode extends AppCompatActivity {
     private TextView date;
     private TextView location;
     private TextView score;
+    private ImageView photo;
     private Button save;
     private ListView commentSection;
     private ArrayList<String> comments;
@@ -27,6 +39,9 @@ public class ViewQRCode extends AppCompatActivity {
     private EditText commentInput;
     private Button postComment;
     private QRCode qrcode;
+    private FirebaseStorage storage;
+    private StorageReference storageReference;
+    private StorageReference photoRef;
 
     /**
      * This method creates the inital interface and obtains the necessary permissions
@@ -36,7 +51,7 @@ public class ViewQRCode extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_view_qrcode);
-        qrcode = getIntent().getParcelableExtra("qrcode_info");
+        qrcode = getIntent().getParcelableExtra("qrcode_info2");
         // Set variable data
         title = findViewById(R.id.title);
         hashedID = findViewById(R.id.hashedID_text);
@@ -47,6 +62,9 @@ public class ViewQRCode extends AppCompatActivity {
         commentSection = findViewById(R.id.comments);
         commentInput = findViewById(R.id.comment_editor);
         postComment = findViewById(R.id.submit_comment);
+
+        // display photo
+        displayPhotoFromFirebase();
 
         // Set values
         hashedID.setText("Hashed ID: " + qrcode.getId());
@@ -75,6 +93,39 @@ public class ViewQRCode extends AppCompatActivity {
         commentSection.setAdapter(commentAdapter);
 
 
+    }
+    /**
+     * This method displays the QR code photo if there is a photo associated to it stored
+     * in Firebase.
+     */
+    public void displayPhotoFromFirebase() {
+        storage = FirebaseStorage.getInstance();
+        storageReference = storage.getReference();
+        String path = SingletonPlayer.player.getUsername() + "/" + qrcode.getImageIDString();
+        photoRef = storageReference.child(path);
+
+        File localFile = null;
+        try {
+            localFile = File.createTempFile("images", "jpg");
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        File finalLocalFile = localFile;
+        photoRef.getFile(localFile).addOnSuccessListener(new OnSuccessListener<FileDownloadTask.TaskSnapshot>() {
+            @Override
+            public void onSuccess(FileDownloadTask.TaskSnapshot taskSnapshot) {
+                // Local temp file has been created
+                Uri imageUri = Uri.fromFile(finalLocalFile);
+                photo.setImageURI(imageUri);
+            }
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception exception) {
+                // Handle any errors
+            }
+        });
+        photo = findViewById(R.id.qr_image);
     }
 
     /**
