--- QRGameTeam15/app/src/main/res/layout/activity_take_photo.xml
@@ -1,33 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<!-- updated by emily with take photo code (START)-->
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".TakePhoto">
-
-
-    <Button
-        android:id="@+id/button_takePhotoUpload"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:gravity="center"
-        android:text="Take Photo and Upload"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/imageView_capturePhoto"
-        android:layout_width="406dp"
-        android:layout_height="620dp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/button_takePhotoUpload"
-        app:srcCompat="@drawable/ic_launcher_background" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
-<!-- updated by emily with take photo code (END)-->
