+++ QRGameTeam15/app/src/main/res/layout/activity_view_qrcode.xml
@@ -9,7 +9,6 @@
 
     <ImageView
         android:id="@+id/qr_image"
-        android:src="@drawable/practice"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:adjustViewBounds="true"
